put the downloaded videos in a new folder
instead of putting the downloaded videos and audio in the playlists folder, I create a new folder for all the downloaded videos according to the playlist name. For example, cnn_abridged.json playlist will be downloaded in the /playlist/cnn_abridged folder. If the folder does not exist, it will be created. If the folder already exists, the folder is deleted and the videos are re-downloaded.